+++ tests/checkbox/checkbox-provider-tdx/units/tests/jobs.pxu
@@ -131,6 +131,18 @@ command:
   export LD_LIBRARY_PATH=/usr/lib/x86_64-linux-gnu/:$LD_LIBRARY_PATH
   setup-env-and-run test_guest_measurement.py
 
+id: tdx-guest/td-ita
+category_id: tdx-guest
+flags: simple
+_summary: Test guest Intel Trust Authority
+depends:
+after:
+requires:
+  executable.name == 'qemu-system-x86_64'
+command:
+  export LD_LIBRARY_PATH=/usr/lib/x86_64-linux-gnu/:$LD_LIBRARY_PATH
+  setup-env-and-run test_guest_ita.py
+
 id: tdx-guest/td-guest-reboot
 category_id: tdx-guest
 flags: simple
